Mouse movemovemnt/zoom working

+++ src/coor.c
@@ -6,13 +6,13 @@
 /*   By: zfaria <zfaria@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/27 13:06:52 by zfaria            #+#    #+#             */
-/*   Updated: 2019/04/01 13:53:18 by zfaria           ###   ########.fr       */
+/*   Updated: 2019/04/01 19:16:53 by zfaria           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <fractol.h>
 
-t_coor	*new_coor(int x, int y, int color)
+t_coor	*new_coor(double x, double y, int color)
 {
 	t_coor *coor;
